backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
pydantic==2.8.2
PyJWT==2.9.0
